requirements.txt
42cc-pystyle==0.0.16
Django==1.6.7
Pillow==4.2.1
South==1.0.2
django-appconf==1.0.1
django-image-cropping==0.9.2
easy-thumbnails==2.1
flake8==2.6.2
mccabe==0.5.3
olefile==0.44
pycodestyle==2.0.0
pyflakes==1.2.3
six==1.10.0
